Restructured files + Dans comments
