+++ pkg/pdatatest/pmetrictest/metrics_test.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"go.opentelemetry.io/collector/pdata/pcommon"
 	"go.uber.org/multierr"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/golden"
@@ -441,38 +440,5 @@ func TestCompareMetrics(t *testing.T) {
 				assert.EqualError(t, err, tc.withOptions.Error())
 			}
 		})
-	}
-}
-
-func TestSortAttributes(t *testing.T) {
-	// Create a sample input map
-	mp := pcommon.NewMap()
-	mp.PutStr("name", "John")
-	mp.PutInt("age", 30)
-	mp.PutBool("isStudent", false)
-	subMap := pcommon.NewMap()
-	subMap.PutStr("city", "New York")
-	subMap.CopyTo(mp.PutEmptyMap("location"))
-
-	// Call the sortAttributeMap method
-	sortAttributeMap(mp)
-
-	// Verify the sorted keys
-	expectedKeys := []string{"age", "isStudent", "location", "name"}
-	actualKeys := []string{}
-	mp.Range(func(key string, _ pcommon.Value) bool {
-		actualKeys = append(actualKeys, key)
-		return true
-	})
-
-	// Check if the keys are sorted correctly
-	if len(expectedKeys) != len(actualKeys) {
-		t.Errorf("Incorrect number of keys. Expected: %d, Actual: %d", len(expectedKeys), len(actualKeys))
-	}
-
-	for i, key := range expectedKeys {
-		if key != actualKeys[i] {
-			t.Errorf("Incorrect key at index %d. Expected: %s, Actual: %s", i, key, actualKeys[i])
-		}
 	}
 }
