+++ receiver/httpcheckreceiver/scraper_test.go
@@ -96,7 +96,6 @@ func TestScaperScrape(t *testing.T) {
 			expectedMetricGen: func(t *testing.T) pmetric.Metrics {
 				goldenPath := filepath.Join("testdata", "expected_metrics", "metrics_golden.yaml")
 				expectedMetrics, err := golden.ReadMetrics(goldenPath)
-				// pmetrictest.SortAllAttributes(expectedMetrics)
 				// golden.WriteMetrics(t, goldenPath, expectedMetrics)
 				require.NoError(t, err)
 				return expectedMetrics
@@ -116,7 +115,6 @@ func TestScaperScrape(t *testing.T) {
 			expectedMetricGen: func(t *testing.T) pmetric.Metrics {
 				goldenPath := filepath.Join("testdata", "expected_metrics", "endpoint_404.yaml")
 				expectedMetrics, err := golden.ReadMetrics(goldenPath)
-				// pmetrictest.SortAllAttributes(expectedMetrics)
 				// golden.WriteMetrics(t, goldenPath, expectedMetrics)
 				require.NoError(t, err)
 				return expectedMetrics
@@ -136,7 +134,6 @@ func TestScaperScrape(t *testing.T) {
 			expectedMetricGen: func(t *testing.T) pmetric.Metrics {
 				goldenPath := filepath.Join("testdata", "expected_metrics", "invalid_endpoint.yaml")
 				expectedMetrics, err := golden.ReadMetrics(goldenPath)
-				// pmetrictest.SortAllAttributes(expectedMetrics)
 				// golden.WriteMetrics(t, goldenPath, expectedMetrics)
 				require.NoError(t, err)
 				return expectedMetrics
